added support for running app and storybook
